+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.23] - createdAt, updatedAt on user document
+
+- When user registers, `createdAt` and `updatedAt` will be added to user document.
+- Whenever user updates profile, `updatedAt` will be updated and `userChange` event fires.
+
 ## [0.0.22] - push notification setting change. user language setting.
 
 - change. push notification settings has been changed.